Update spinner to work with v2 of react-transition-group
v2 removes static lifecycle methods on transitionGroup children and
enforces the children of transitionGroup to be transition instead. This
posed some problems with the current structure of Spinner (Parent) -> Glyph (Child)

It looks like transitionGroup was only being used to handle exit transitions
in a child component that relied on exit state from the parent as just using a
plain transition in the child component would result in the child component
re-rendering and preventing either an exit animation.

Now that everything is in one component, we can use a simple Transition
component.