ED-3294: Remove support for unused title attr on link in BB transformer
We were supporting having 'title' as a attr on the link schema node, despite our schema
not supporting it. I removed the unused code and also uncommented skipped tests for the
link, a few of them could be deleted because they were testing the functionality for title
which we don't support :). I also added a reference to a followup action to fix the
escaping for the link node as currently we aren't escaping the href content at all