ED-3294: Fix BB by padding PM contenteditable instead of container
This fixes a very strange bug that presents itself in firefox when dragging
and dropping content into the editor. If the parent of the content editable
has padding, and the user drags a file not onto the content editable but the
padding itself, an InputEvent (rather than a DragEvent) will be triggered and
result in the media skipping any paste/drop handling. If the padding is
around the content editable itself, then everything works as expected.

This edge-case is still slightly present, if you drop the file over the edge
of the content editable padding, however this already affects BB today. As
we are disabling content with Data URI's in a separate commit, this edge case
will be ignored and not cause any impact.

n.b. one way to see if this is occuring, is when drag 'n dropping a file into
the browser, it will display a cursor inside the content editable.