Add withContextAcrossPortal HOC to replace withContextFromProps
The new HOC allows the context to be propagated across portals at a
point closer to the portal boundary, the portal boundary component, rather
than the parent that renders the portal boundary. This is necessary for
chained context pattern used in the Analytics package.
The example has also been updated to demonstrate components accessing
context within portals as it wasn't showcasing that previously