Fix test issues caused by default export component class and test component import mismatch
The codemod will be updated to export the non-HOC wrapped components and
use those in tests so that existing tests do not break. We need the
component import/export names currently used for the default export to match
so that things will work when we change them to a named import/export.
Also, the filepaths for the component under test will need to point to
the component file itself rather than a src/index file as the non-HOC
wrapped component will not be exported by the package itself.