Update calendar to manage controlled/uncontrolled state itself
We found an issue in the new react-ctrl API that was solved by moving
the ownership of mapping default props to initial state from react-ctrl
to the component itself (by setting default props for your defaultXXX
props and then referencing them in the initial state value).
Once this was removed, the only thing react-ctrl was doing was providing
a merged props/state object via function as child and it was decided to
just do this in the component itself for now.

We also found that with using either the function as child version of
react-ctrl or managing mapped state ourselves, we have to replace
this.state access with mapped state anywhere in our component, not just
render. Therefore we have added a getState method to the component to
manage state access that handles mapping state.