Add analytics prop callback docs for avatar, blanket and breadcrumbs
The analytics event map is optional in the flow types to prevent
flow errors. Fortunately, the atlaskit docs do not detect the arg
as optional.
Technically, typing the base component props with this UI event is
incorrect, however, we cannot find an easy way to type the patched
prop callbacks within the analytics HOCs using flow.