ED-4962: Add canInsertLink object to hyperlink state.
This isn't the ideal solution. Ideally, the InsertBlock toolbar
should allow you to get a function that would determine the
disabled / active state of the toolbar button. Consider this
a temporary workaround pending a refactor of the ToolbarInsert
plugin. This is also why InsertBlock is the only consumer of
this new state field