FS-3068 Make 'source' optional in GasCorePayload
In most cases the 'source' field is expected to be set by the integrating
product. The problem with having it defined as mandatory is that components
end up setting it to the default value "unknown" which will take precedence
on any (more correct) value the product might set. Although this depends more
specifically on the behaviour of the relevant AnalyticsListener, today that's
what would happen for the existing ones. They are also already handling the
case where the 'source' field hasn't been set and set the fallback value
"unkown".