Fix cursor and resize chevron flickering while dragging
Setting pointerEvents none on the body caused the rbdnd body styling to
fail. It also caused mouseEnter/mouseLeave events to be triggered on
LayoutManager whenever the cursor moused out of the item being dragged.

To fix this we are now notifying LayoutManager when a drag has started
so it can set some nav specific styles for us.