Prevent unnecessary re-renders when starting a drag
Ending a drag will cause a re-render of the whole nav since the renderer
is passed a fresh set of items when updating a view so we cannot fix
that currently. The nav will at least not re-render when ending a drag
in its original location.

Also rename itemIsDragging state variable and align
ContainerNavigationMask to use disableInteraction prop