Rename SortableSection to SortableContext to prevent nested sections
We needed to nest SortableSection inside a MenuSection component for
nested transitions to animate since they rely on the same component
being used in both the parent and child view. Using a SortableSection
inside MenuSection added extra padding to items and broke the semantic
rule of sections where they should not be nested.

SortableContext now only renders a DragDropContext component and does
not render a section.